[Driver][SYCL] Improve diagnostic behaviors for intel_gpu targets
When using multiple targets, like
-fsycl-targets=intel_gpu_pvc,intel_gpu_bdw, we were emitting a
diagnostic that the latter value was being ignored as the values were
similar.  This is not a correct diagnostic, as the values have identical
triples, but not an identical arch target.

Update how these diagnostics are produced, depending on using the old or
the new offloading model.