[Bench] Fix validation standards for compare.py (#19689)
Due to the performance BMG machine being taken down for intel/llvm
recently, there is a gap in the data stored in
intel/llvm-ci-perf-results. In preparation for bringing up the BMG
machine next week, I have swapped out the data for the data from
unified-runtime CI.

In order for the data in unified-runtime CI to work, we need to relax
the validation standards in compare.py.