[SYCL][NFCI] Move abs and div to cstdlib header (#19671)
The C++ specification defines that `abs(<integer>)` and `div` functions
are provided by `cstdlib`.
Whether they're available or not after `#include <cmath>` is an
implementation detail of the underlying C++ standard library
implementation.

With this change we move the definition to the correct wrapper file
aligned with the C++ specification.