[SYCL][NATIVECPU] Fix TargetInfo initialization (DataLayoutString) (#19711)
Some clang targets (for example for Arm) may not initialize
`DataLayoutString` in their TargetInfo constructor, which meant calls
`getDataLayoutString` right after the constructor cause an assert.

This PR defers the initialization of the NativeCPU TargetInfo to after
the call of `handleTargetFeatures` of the host cpu Target info, which on
Arm initializes `DataLayoutString`, but may also catch more
initializations.