[SYCL][NativeCPU][CMake] Simplify OCK CMake handling (#19704)
I'm trying to make this repo more easily packaged by Linux distros, and
OCK came up in some feedback from NixOS (see linked issue).

Overall, the external dependency behavior should work as follows:

1) Use the user specified directory if given
2) If not, search the system for the dependency's install location and
use that
3) If not found, fetch it

OCK seems a bit special because installing OCK locally doesn't seem to
install any CMake files nor does it install the files required to use it
to build NativeCPU support. Seems we actually need the git tree, not the
installed result of a built OCK. Strange but let's go with it.

You can already provide a precreated clone to use to build with
`FetchContent` using `FETCHCONTENT_SOURCE_DIR_<uppercaseName>`, so there
is no need for variable to set the directory and a custom repo and tag.

Remove variables that duplicate functionality that already exists in
`FetchContent`, and add a simple search for the source directory on the
system.

Issue: https://github.com/intel/llvm/issues/19635

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>