[SYCL][NATIVECPU] Port ock compiler internals into sycl native cpu
It is desirable to decouple native cpu from needing to fetch the external
oneAPI Construction Kit to allow it to be maintained more smoothly.

To allow a smooth port, much of the files have been copied verbatim, but with
a different use of cmake to git in with llvm components. There is a great
deal of tidying up that will be required after this port to fit better
in with LLVM and prune what is not necessary.

Some documentation has been added, but this will be built upon in future
commits.