[SYCL] Free function kernels bugfix (#19535)
When free function kernel is a function template which in turn has
arguments that are also templates, we may not print the latter properly
into integration header.

Specifically, if an argument of a templated free function kernel has an
`enum` template argument of its own then namespace qualifiers for it
will be missing.

This patch fixed that.