[libc++] Fix return type of ilogb(double) (#19755)
This commit addresses a seemingly unintentional return type of the ilogb
overload taking a double. Currently it returns double, while it is
supposed to return int.

The return types seems to be covered by
libcxx/test/std/numerics/c.math/cmath.pass.cpp, but the issue would only
show up if we tested with a libc that doesn't provide the ilogb(double)
overload, which we don't.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>