[CI] Enable UR runs in benchmarking CI (#19753)
Temporary change enabling UR benchmarking.

This is not the most ideal solution, but currently no benchmarking data
is being collected for the UR; this data is needed urgently. A proper
solution can be discussed after UR benchmarking has been re-enabled.

**Note:** The new benchmark CI runs are using a docker environment,
there seems to be a very consistent performance penalty of 2
microseconds as a result compared to previous UR benchmark CI runs that
are not using a docker environment. Please see
https://intel.github.io/llvm/benchmarks/?tags=UR&runs=Baseline_PVC_L0%2CCommit5661e7275b64a6c00a96ae295815f40a9384e958_PVC_L0
: this compares performance of latest commit with commit 5661e727, which
was the commit used to benchmark the last run result from July 30th.

I am currently looking into ways to reduce this penalty, for now I am
opening this PR as the data is still useful. It's worth noting that SYCL
only incurred 0.5-1 microsecond penalty from this move, so there may be
room for improvement here.