[SYCL] Fix reported P2P accessibility across platforms
The current P2P extension implementation assumes that the devices are
from the same platform, which may not be the case. In cases where they
are not the same, the adapters will try to access the handles as if they
were the layout of the corresponding adapter, which may manifest as
invalid accesses or unexpected read internal handle values, which can in
turn conflict with actual device handles, potentially causing false
positives. This commit adds a check that the devices are from the same
platform before requesting the information from the backend.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>