[UR][CMake] Improve support for pre-installed Level Zero and simplify (#19763)
First, try finding preinstalled Level Zero using pkg-config first
If that fails, try to find include/header dirs on the OS
And if those both fail, fetch from github (both headers and loader).

No need for variables to set the repo and commit to clone, it's highly
unlikely anyone will do this and if they want to they can use the
existing CMake FetchContent override infrastructure.

I also removed a variable about whether to use fetchcontent or not. We
should just do what everyone expects by default, try to find it on the
system and if we do use it, otherwise fallback to fetching it.

This was requested by NixOS developers, see the below issue.

https://github.com/intel/llvm/issues/19635

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>