[SYCL] Fix linkage adjustment of kernels (#19771)
This patch resolves #19409

Two main changes were done:

1. Corrected method of detection that kernel is defined as an inline
function. The original code detecting this was introduced in #338, but
by some reason we looked into a headers-provided wrapper that calls a
kernel instead of a kernel itself. Alongside with fixing this, a
dedicated clang-level test was added to check this behavior

2. Fixed a bug introduced by incorrect conflict resolution (commit
fcd95a956bccc5aaa96c8bee5ced64f4cdd999e2) with llvm/llvm-project#137882.
Specifically, when deciding if a function (that is expected to be a SYCL
kernel) linkage should be promoted to non-discardable we used to check
for OpenCL kernel attribute. That attribute is not present in SYCL
headers and therefore was only generated for actual SYCL kernels emitted
by front-end. However, with kernel attributes unified, we now promote
linkage of non-kernel wrapper function that has `sycl_kernel` attribute
attached to it explicitly by SYCL headers. The fix here is to only react
to the kernel attribute if it was added implicitly, to preserve original
behavior.