[UR][Offload] Implement (most) kernel info queries
Almost all of the required kernel info queries have been implemented.
The only exception is `NUM_ARGS`, which isn't reported by liboffload,
CUDA or HSA (the UR plugins return the *current* number of arguments,
not the total). This test has been marked as a known failure and tests
that use it as a placeholder value have been changed to use NAME.