[SYCL] fix for flaky ~event failure in Win unit tests (#19762)
Windows shutdown/teardown doesn't have the exact same timing as on
Linux, and even more in the the unit tests which are statically compiled
( so no DllMain() call, which means no space between the calls to
`shutdown_early()` and `shutdown_late()` ). Here we remove a race
between the platform deletion in the mock and the win shutdown. This
fixes a flaky failure we are seeing in a couple of the unit tests.

---------

Signed-off-by: Chris Perkins <chris.perkins@intel.com>