[SYCL] Disable dead arg elimination for free function kernels (#19776)
This PR disables dead argument elimination for free function kernels.
When using a free function kernel, the user sets the arguments manually
using `handler::set_arg` but if certain arguments are optimized away in
case they are not used, the implementation needs to be aware of this and
remove the relevant `set_arg` calls. Instead, its much more feasible to
just disable dead arg elimination in this case.

---------

Co-authored-by: Alexey Sachkov <alexey.sachkov@intel.com>