[CI] Sign nightly linux build
This PR adds an optional step in sycl-linux-build to sign a build using
the sigstore/cosign action.

We should sign our release binaries. Currently we're performing it
manually outside CI due to security reasons. But in future we'd like to
switch to cosign instead. Therefore we're introducing this as part of
nightly for now to test it in advance.
Cosign insures that a blob was signed exactly by the specific workflow.

Test run: https://github.com/intel/llvm/actions/runs/17404358377
To verify:
cosign-windows-amd64.exe verify-blob \
--bundle /path_to/sycl_linux.tar.gz.sigstore.json /path_to/sycl_linux.tar.gz \
--certificate-identity https://github.com/intel/llvm/.github/workflows/sycl-linux-build.yml@refs/heads/BRANCH_NAME