[SYCL][CMake] Simplify emhash fetch code (#19894)
Upstream added a CMake install target, so now we can rely on that when
using a system installed version, and we can create an interface target
to emulate the real one when using FetchContent.

The upstream CMake install target installs all files directly in the
given prefix folder, so with `CMAKE_PREFIX_PATH=/usr` it will install
`hash_table8.hpp` directly to `/usr`, not in an `emhash` subfolder.

Since it seems that's how upstream wants this packaged, update our
includes to remove the fake `emhash` subdirectory.

Issue: https://github.com/intel/llvm/issues/19635

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>