[vecz] Avoid mishandling poison.
We were using 'and' to filter branch conditions, trusting that if an
entry mask had it set to false, it would assuredly be kept false. This
does not work when a branch condition is computed as 'poison', as 'false
and poison' evaluates to 'poison', not 'false'.

We can generate 'select' instructions instead, which avoid this problem.