[NATIVE_CPU][SYCL] Switch to using native_cpu compiler pipeline inline
This brings the native cpu compiler pipeline files directly under the ownership
of intel/llvm. This removes the direct dependence on the oneAPI Construction
Kit, although the history of those files still exists under intel/llvm and the
originals still exist at
https://github.com/uxlfoundation/oneapi-construction-kit.

This is the post merge update to the the oneAPI Construction Kit move of the compiler
pipeline files for Native CPU and replaces the original FetchContent method.

This updates the CMakeLists.txt to be appropriate for llvm components,
rather in the style of oneAPI Construction Kit. It also applies clang-format-20 to all the
.h and .cpp files brought in. Other than that the .cpp and .h files are unchanged.

A brief explanation and limitations of the importing of the files are under l
lvm/lib/SYCLNativeCPUUtils/compiler_passes/compiler_passes.rst.

It also updates the NATIVECPU_USE_OCK to be on in sycl-linux-build.yml.