[SYCL] Kernel submit wrapper functions for unit tests (#19884)
The new kernel submit wrapper functions allow to select the submission
function type between a standard handler-based function and queue
shortcut function. This change allows to increase the test coverage and
allows for testing no-handler submission path using existing unit tests.
This PR introduces the single_task submission function wrappers, and the
use of the wrappers in a single unit tests set, but the goal is to
extend the set of wrappers and their usage in unit tests, in the future.