[SYCL] remove -### check in findLibspirvPath (#19784)
In our downstream repo, bf21b9de88b3 exposed a regression that libspirv
path is invalid in -### output. Before bf21b9de88b3, the path was
correctly found by `searchAt(D.ResourceDir, "..", "..", "clc")`, which
was deleted by the commit.
This PR deletes -### check to align behavior with when -### is not
present.

---------

Co-authored-by: Mészáros Gergely <maetveis@gmail.com>