[SYCL][CMake] Fix emhash build error (#19967)
We only need to link against `emhash` in build, it's an interface
library representing a header-only project (so all linking does is add
includes), and it's used internally in `xptifw` source files and in a
single non-installed header, but some other part of the code could in
theory use that non-installed header during build, so we need to link
against it, but only for build.


Manually checked the generated CMake code.


Issue introduced in https://github.com/intel/llvm/pull/19894

Example
[error](https://github.com/intel/llvm/actions/runs/17439817298/job/49520655604?pr=19963):

```
CMake Error: install(EXPORT "unified-runtime-targets" ...) includes target "xptifw" which requires target "emhash" that is not in any export set.
```

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>