[SYCL] Event-less APIs synchronization with the scheduler
Commands submitted to the scheduler need to unconditionally
be associated with an event (for both event and event-less APIs).
This is because some commands might already be scheduled
and waiting for the submission, and a newly submitted command
need to return an event which can be used by the in-order type
queue to properly order the commands.