[SYCL][Doc] Remove incorrect "KernelName" in spec
The overloads of `parallel_for` and `nd_launch` that take a
`sycl::kernel` object should not allow the application to specify a
name for the kernel.  The kernel which is invoked by these overloads
is determined at runtime (by the value of the `sycl::kernel` object),
so it's not possible to assign a static name.

The implementation was already correct, so this is just fixing a
documentation bug.