[SYCL] Propagate data from `handler` to `DeviceKernelInfo` when using old binaries (#19993)
`DeviceKernelInfoPtr` wasn't/isn't used a single source of information
neither before not after this PR, so things weren't "broken". However,
we want to change that soon and for that we need all the necessary
information from `handler`/`handler_impl` to be available in
`DeviceKernelInfoPtr`. That change brings us closer to that state.