[SYCL] fix for __sycl_unregister_lib() on Windows and tests (#19633)
`__sycl_unregister_lib()` is not being called on Windows when using
shared libraries due to a limitation in LLVM/clang To work around this,
on Windows we register both `__sycl_register_lib()` and an `std::atexit`
handler that will call `__sycl_unregister_lib()` . Further, it was
discovered that freeing of the device globals during device images
destruction is duplicate and unnecessary. The `~context_impl` destructor
handles that (and handles it correctly, because a context is needed to
free USM memory). So we remove the unneeded duplication. Adding a test
that stresses `__sycl_unregister_lib()` and makes sure there are no
resource leaks

---------

Signed-off-by: Chris Perkins <chris.perkins@intel.com>