[SYCL] Get rid of device kernel info duplication
With the introduction of DeviceKernelInfo, assert usage and implicit local
argument information is now duplicated in program manager. This patch removes
the duplicate maps and makes it so that device kernel info map is filled out
during image registration, with the compile time information added when it's
available (during the first submission of the kernel).