[SYCL] Remove `UNSUPPORTED` tags for 3 E2E tests (#20021)
These tests have been reported to be failing in
https://github.com/intel/llvm/issues/17305, but they are actually
passing with the latest version of GPU driver `25.31.34666.3`. The
previous failure is most likely caused by running these tests on an
older version of GPU driver (namely `25.05.32567.17`) which doesn't
support ptl -- as you can see in the error output section on that GitHub
issue page, the error message is "Build failed for : ptl-h-a0 with error
code: -11". Therefore, creating this PR to remove the `UNSUPPORTED` tags
for these 3 E2E tests.

Signed-off-by: Hu, Peisen <peisen.hu@intel.com>