[SYCL][NativeCPU] More --shared-libs fixes. (#20081)
* SYCLNativeCPUUtils/CMakeLists.txt: remove EXCLUDE_FROM_ALL. The
rationale for setting EXCLUDE_FROM_ALL was to ensure that users do not
get the veczc binary unnecessarily but that is already handled
differently now that we no longer fetch external sources, and
EXCLUDE_FROM_ALL had the side effect of preventing component libraries
from being installed.
* vecz/tools/CMakeLists.txt: change veczc from tool to utility. The
difference is that tools get installed, utilities by default don't, and
this is not something that is meant to be installed.
* lib/CMakeLists.txt: change include order. SYCLNativeCPUUtils needs to
be included after TargetParser so that veczc correctly links to it.

Fixes #20077