[NFC][SYCL] Prefer `(void)` over `std::ignore` (#20088)
One of my recent PRs to reduce header dependencies of
`id/item/range/etc.` has caused some compile time errors downstream with
`std::ignore` not being declared because we use different version of STL
than the one in the upstream CI.

We do want to have our device headers as lightweight as possible
(especially in the context of SYCL RTC), so it makes sense to just use
`(void)` uniformly instead of `std::ignore`. UR experience also found
that the compile-time price of `(void)` is smaller as well.