[SYCL] Use correct units when specifying image offset and extent (#19936)
`urBindlessImagesImageCopyExp` documents its argument that describe copy
region as using `bytes-pixel-pixel` format, while we were mistakenly
passing `pixel-pixel-pixel` format in there.

This PR adjusts that. It is an `NFC` change for CUDA & HIP adapters,
because for them it only changes the place where pixel to byte
conversion happens, but the change is tricker for L0 adapter. In L0,
there are APIs which expect `pixel-pixel-pixel` format and there are
APIs which accept `bytes-pixel-pixel` format, so some extra care is
needed there.

---------

Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>