[UR][XPTI] Fix RelWithDebInfo and MinSizeRel builds on Windows. (#20078)
CMake `RelWithDebInfo` builds failed on Windows because the Unified
Runtime attempted to link with a `Debug` build of XPTI. Fortunately, the
mismatch was detected by `lld-link` and it failed the build with the
following error:
`error: /failifmismatch: mismatch detected for '_ITERATOR_DEBUG_LEVEL'`
The mismatch was due to a CMake generator expression in the Unified
Runtime `CMakeLists.txt` that failed to account for `RelWithDebInfo` and
`MinSizeRel` builds.

Investigation of the above also resulted in the identification of an
issue in the XPTI `CMakeLists.txt` file; use of
`CMAKE_CXX_FLAGS_MINSIZEREL` for a `RelWithDebInfo` build and
`CMAKE_CXX_FLAGS_RELWITHDEBINFO` for a `MinSizeRel` build.