[SYCL][E2E][NATIVE_CPU] Add Native CPU e2e PR testing
The Native CPU work is being passed onto a team in Intel now that
all of Native CPU is independent of the oneAPI Construction Kit.

Native CPU does not pass all of E2E testing, either through bugs or
unimplemented functionality. All are currently being marked as
UNSUPPORTED until these failures are processed and tracked by the new
team.

This also adds a separate job for E2E build and run for Native CPU,
as it cannot be added to the prebuilt due to build fails.