[SYCL][NATIVE_CPU] Update docs for Native CPU compiler pipeline (#20042)
This integrates the appropriate compiler documentation originally in the
oneAPI Construction Kit (OCK) into the NativeCPU compiler pipeline
documenation.

It has been updated to try to reflect the Native CPU pipeline, and
remove some of the references to OCK's structures, as well as moving
some of the documentation to markdown files to be consistent with some
of the other documentation.

Some of it may be irrelevant for Native CPU, and if so this should be
updated over time.

Support was added for the mermaid flowcharts in the config.