[UR] Make offload program CTS tests pass (#19948)
The CTS tests for the program part of the API now pass (or are skipped).
This mostly was a matter of implementing olProgramGetBuildInfo and
olProgramLink (but only for one binary - specifying more than one is an
error). The build failure tests have also been tweaked a bit to better
reflect their optional nature.