[SYCL] Fix race during image decompression  (#20124)
This is a cherry-pick of intel/llvm#19981

**Problem**
There can be race during image decompression. Consider the case when one
thread is reading the decompressed buffer while another thread modifies
it. This can lead to invalid SPIRV errors emitted by IGC.

**Proposed solution**
Use `std::call_once` to ensure that only one thread does the
decompression of an image.

Patch-by: Udit Kumar Agarwal <udit.agarwal@intel.com>