[SYCL][Native CPU] Avoid dangling references.
In multiple places, we take a reference to an element in LoopMasks and
expect that reference to remain valid. LoopMasks is a DenseMap which does
not promise that references to existing elements remain valid as new
elements are inserted. We pre-populate LoopMasks to try and avoid this
becoming a problem, but we were pre-populating it only with top-level
loops, when sub-loops may also cause references to become invalid. Adjust
the pre-population to take all loops into account.