[SYCL][XPTI] Performance improvements for performance oriented toolchains (#19953)
Implementation of XPTI within SYCL combined debug and performance
streams over time and this PR improves the performance by 15-20% for
toolschains by separating the debug data from performance streams.
- Limits the amount of debug information that is attached to the default
"sycl" performance stream.
- If addition metadata is needed, toolchains will have to subscribe to
"sycl.debug" stream
- Unified runtime created a new trace event for each function call which
adds significant overhead to each call when a collector is enabled. This
is now minimized by using a global event for all of UR API calls.
- CUDA plugin implementation of XPTI events enabled both performance and
debug streams all the time, irrespective of where there was a tool
subscribing to the data.
- If tools want good performance then they need to subscribe to "sycl"
stream. If they want full information and performance is not as
important then subscribing to "sycl.debug" stream is the right choice.

---------

Signed-off-by: Vasanth Tovinkere <vasanth.tovinkere@intel.com>
Co-authored-by: Gainullin, Artur <artur.gainullin@intel.com>