[SYCL] Reduce scope of the `SYCL2020_DISABLE_DEPRECATION_WARNINGS` macro (#20155)
We have two kinds of deprecations in our implementation:
1. Functionality that is deprecated by the SYCL 2020 specification. It
must still be supported and will only be removed in the future version
of the standard. We can only remove it once we drop support for the SYCL
2020 standard all together;
2. Functionality that is deprecated by us (some non-standard API
accidentally made public for whichever reason, changes to extensions,
etc.). Such functionality can be removed by us way quicker and it isn't
tied to the SYCL 2020 standard;

By some reason, the macro which is intended to control warnings about
the group (1) also controlled warnings about the group (2) and this
patch fixes that.

`SYCL2020_DISABLE_DEPRECATION_WARNINGS` changed it so that it **only**
control warnings from the group (1).

Warnings from the group (2) can only be disabled with a newly introduced
internal macro `__SYCL_DISABLE_DEPRECATION_WARNINGS` which we need to
reduce the noise from building `libsycl.so` and unit-tests for it which
still can reference deprecated functionality from the group (2).

There is no user-visible macro provided for disabling warnings from the
group (2). Anyone who wants to silence those warnings should use regular
`-Wno-deprecated-declarations`.