[SYCL][E2E][NATIVE_CPU] Add Native CPU e2e PR testing (#19961)
The Native CPU work is being passed onto a team in Intel now that
all of Native CPU is independent of the oneAPI Construction Kit.

Native CPU does not pass all of E2E testing, either through bugs or
unimplemented functionality. Most are currently being marked as
UNSUPPORTED until these failures are processed and tracked by the new
team, with a smaller number marked as XFAIL. These are being tracked
under https://github.com/intel/llvm/issues/20142.

This also adds a separate job for E2E build and run for Native CPU,
as it cannot be added to the prebuilt due to build fails.