[SYCL][Native CPU][NFC] Increase safety.
Since we rely on LoopMasks[...] not creating new map entries, since they
should have been created in advance, use LoopMasks.at(...) instead. This
will ensure that if new changes are introduced that result in any loop
masks being missing, it will fail reliably and be much easier to debug.

This commit fixes one issue with BOSCC, but this is still NFC as at the
moment, BOSCC is never used.