[SYCL] Fix Lambda Mangling in Namespace-Scope Variable Initializers. (#20176)
The community counterpart to this PR can be found here:
https://github.com/llvm/llvm-project/pull/159115.
In that PR, we propose a fix for the lambda mangling issue in
namespace-scope variable initializers. This fix has uncovered a number
of related problems. @tahonermann has identified several scenarios where
the lambda mangling is either incorrect or diverges from `GCC`’s
behavior.
While the current patch addresses a high-priority issue flagged in our
downstream compiler, it requires additional time and effort to
thoroughly address the full range of scenarios—some of which are
partially listed in https://github.com/llvm/llvm-project/pull/159115.
The current change successfully fixes the issue affecting our downstream
compiler.
Therefore, we propose to “cherry-pick” the community PR now, while
continuing work in the community repository to complete the fix.

For convenience here is the description of
https://github.com/llvm/llvm-project/pull/159115:
This PR addresses an issue with the mangling of lambdas used as
initializers for global variables within namespaces. According to the
Itanium C++ ABI, lambdas should be uniquely mangled based on their
context. GCC correctly includes the name of the declared variable in the
mangling context for such lambdas, avoiding the need for discriminators
since each lambda is scoped to its respective variable (see
https://godbolt.org/z/38Y8qvvj3).

When C++ code is compiled without CUDA, HIP, or SYCL enabled, lambdas
that don't require external linkage are given internal linkage and
mangled with a $ in their name. When CUDA, HIP, or SYCL support is
enabled, separate compilation for host and device requires that the same
mangled names are observed, even for symbols that have internal linkage.
Code to use external mangling was already present, but the mangled names
for lambdas were incorrectly generated in some cases:

Lambdas in the initializers of global variables used the enclosing
namespace as the mangling context rather than the global variable.
When such global variables were declared in different partial namespace
definitions, discriminators were allocated in the context of a specific
partial namespace definition rather than in the primary namespace
definition leading to mangled name clashes, some of which provoked an
error, some of which resulted in an additional .<n> disambiguator being
silently added to the symbol.
This PR ensures that lambdas used as global variable initializers are
mangled in the context of the declared variable and that a canonical
namespace is used for allocation of mangling discriminators.